api/requirements.txt
pyserial
flask
flask-cors
gunicorn
